+++ CHANGELOG.md
@@ -26,6 +26,7 @@
 - [11344](https://github.com/vegaprotocol/vega/issues/11344) - Include derived parties in accounts API
 - [11268](https://github.com/vegaprotocol/vega/issues/11268) - Include derived parties vesting stats API
 - [11266](https://github.com/vegaprotocol/vega/issues/11266) - Include derived parties rewards API
+- [11357](https://github.com/vegaprotocol/vega/issues/11357) - Support historical game scores
 
 ### 🐛 Fixes